requirements.txt
flask
redis
opencv-python
pillow
numpy
celery
eventlet
flower